refactor!(postgres): allow string types for port
The port was previously specified using the port type, but this does not
allow for any flexibility for the port by the end-user. This means that
running the service in multiple projects would have conflicts if the
port wasn't unique among them. By allowing for a string type, the port
can be controlled by an environment variable and therefore it is
possible to use the service multiple times without everyone agreeing on
unique ports.